refactor(postgres): simplify configmap creation and references
The configmap creation for postInitSQL and postInitApp has been moved
to the end of the template to avoid duplication. The configmapRefs
now use the cmName variable directly, instead of using the default
function again. This makes the code more readable and maintainable.
Also create configmaps for postInitSQL and postInitApp if needed.